modules/image/misc: fix function casts
The return value of a function should only be casted to a compatible
type; lots of "(int) floor(x)" can be simply changed to the implicit
casting rules.

+++ modules/src/image/misc/disocclusion.c
@@ -1563,12 +1563,12 @@ static void geodesic(jordan *jc, int x)
 	    fpivots[(npivots-1)*4+2]=gray2;break;
 	  default:prerror("Probleme de direction");break;
 	  }
-	gx1 = (int) floor(2 * geodPoints[n / 2].ps[0].y + .5);
-	gy1 = (int) floor(2 * geodPoints[n / 2].ps[0].x + .5);
+	gx1 = floor(2 * geodPoints[n / 2].ps[0].y + .5);
+	gy1 = floor(2 * geodPoints[n / 2].ps[0].x + .5);
 	for (pi = 1; pi < geodPoints[n/2].pn; pi++)
 	  {
-	    gx2 = (int) floor(2 * geodPoints[n / 2].ps[pi].y + .5);
-	    gy2 = (int) floor(2 * geodPoints[n / 2].ps[pi].x + .5);
+	    gx2 = floor(2 * geodPoints[n / 2].ps[pi].y + .5);
+	    gy2 = floor(2 * geodPoints[n / 2].ps[pi].x + .5);
 	    drawgeod(gx1,gy1,gx2,gy2,gray1,gray2,pi-1);
 	    gx1=gx2;gy1=gy2;
 	  }
@@ -1578,8 +1578,8 @@ static void geodesic(jordan *jc, int x)
 	   want to avoid conflicts between gray values */
 	for (pi = 0; pi < geodPoints[n/2].pn; pi++)
 	  {
-	    pos = (int) floor(geodPoints[n / 2].ps[pi].x)
-		 * col_number + (int) floor(geodPoints[n / 2].ps[pi].y);
+	    pos = floor(geodPoints[n / 2].ps[pi].x)
+		 * col_number + floor(geodPoints[n / 2].ps[pi].y);
 	    ptrpiv=pivots+8*pi;
 	    npi1=0;
 	    while (*ptrpiv!=1) {npi1++;ptrpiv++;}
